backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
SQLAlchemy
python-jose
python-jose[cryptography]
passlib[bcrypt]
python-multipart
bcrypt==4.0.1
python-dotenv
pydantic-settings
pydantic[email]